+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## TO BE RELEASED
 
+* OPC-139 LTI Oauth config change from Upstream
 * MI-164: set nginx logrotate to only keep 30 
 * MI-171: nginx config reload needs to watch for changes to both the ssl key *and* cert
 
@@ -29,7 +30,6 @@
 +                  <queue physicalName="SCHEDULER.CaptureAgentSync" />
                    <queue physicalName="SCHEDULER.Liveschedule" />
                  </forwardTo>
-
 
 ## v2.7.1 - 04/26/2019